requirements.txt
Django==4.2.2
django-environ==0.9.0
djangorestframework==3.14.0